Add class for MOC streamfunction
