+++ mpas_analysis/ocean/__init__.py
@@ -3,4 +3,4 @@ from climatology_map import ClimatologyMapSST, ClimatologyMapMLD, \
 from time_series_ohc import TimeSeriesOHC
 from time_series_sst import TimeSeriesSST
 from index_nino34 import IndexNino34
-# from meridional_overturning_circulation import moc_streamfunction
+from streamfunction_moc import StreamfunctionMOC
